Intensive testing of LabelingDistributionCalculator
Former-commit-id: 552a073f276c050e9a7266fd66076a8c6019f017
Former-commit-id: 5fc271992951ef5d98baeadc21fd3887c5dc29c3

+++ DeconTools.UnitTesting2/Utilities/IsotopicProfileElutionExtractorTests.cs
@@ -26,7 +26,7 @@ namespace DeconTools.UnitTesting2.Utilities
             double maxMZ = 754;
 
             int[] scans;
-            int[] intensities;
+            float[] intensities;
             double[] mzBinVals;
